Update day1_joke_teller.py
Project Name: Day 1 – Joke Teller

This is my first Python mini-project where I practiced the basics I learned on Day 1:

• Using modules

• Writing comments

• Taking user input

• Working with variables

The project asks for the user’s name and age, then uses different modules to make the program more fun. It prints a joke using the pyjokes module, adds an emoji with the emoji module, and even reads the joke aloud with pyttsx3.

The goal of this project was not just to write code, but to actually create something small and interactive with what I learned on the first day.